refactor: IndexPageGenerator 인터페이스를 도입한다
